create own class for injecting preview urls into commands
