+++ infrastructures/openshift/src/main/java/org/eclipse/che/workspace/infrastructure/openshift/OpenShiftInternalRuntime.java
@@ -15,14 +15,11 @@ import com.google.inject.assistedinject.Assisted;
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.api.model.Service;
-import io.fabric8.kubernetes.api.model.ServicePort;
 import io.fabric8.openshift.api.model.Route;
-import io.fabric8.openshift.api.model.RouteSpec;
 import io.opentracing.Tracer;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 import javax.inject.Inject;
 import javax.inject.Named;
@@ -42,11 +39,11 @@ import org.eclipse.che.workspace.infrastructure.kubernetes.cache.KubernetesRunti
 import org.eclipse.che.workspace.infrastructure.kubernetes.namespace.pvc.WorkspaceVolumesStrategy;
 import org.eclipse.che.workspace.infrastructure.kubernetes.util.KubernetesSharedPool;
 import org.eclipse.che.workspace.infrastructure.kubernetes.util.RuntimeEventsPublisher;
-import org.eclipse.che.workspace.infrastructure.kubernetes.util.Services;
 import org.eclipse.che.workspace.infrastructure.kubernetes.util.UnrecoverablePodEventListenerFactory;
 import org.eclipse.che.workspace.infrastructure.kubernetes.wsplugins.SidecarToolingProvisioner;
 import org.eclipse.che.workspace.infrastructure.openshift.environment.OpenShiftEnvironment;
 import org.eclipse.che.workspace.infrastructure.openshift.project.OpenShiftProject;
+import org.eclipse.che.workspace.infrastructure.openshift.provision.OpenShiftPreviewUrlCommandProvisioner;
 import org.eclipse.che.workspace.infrastructure.openshift.server.OpenShiftServerResolver;
 
 /**
@@ -76,6 +73,7 @@ public class OpenShiftInternalRuntime extends KubernetesInternalRuntime<OpenShif
       OpenShiftEnvironmentProvisioner kubernetesEnvironmentProvisioner,
       SidecarToolingProvisioner<OpenShiftEnvironment> toolingProvisioner,
       RuntimeHangingDetector runtimeHangingDetector,
+      OpenShiftPreviewUrlCommandProvisioner previewUrlCommandProvisioner,
       Tracer tracer,
       @Assisted OpenShiftRuntimeContext context,
       @Assisted OpenShiftProject project) {
@@ -98,6 +96,7 @@ public class OpenShiftInternalRuntime extends KubernetesInternalRuntime<OpenShif
         toolingProvisioner,
         null,
         runtimeHangingDetector,
+        previewUrlCommandProvisioner,
         tracer,
         context,
         project);
@@ -165,23 +164,5 @@ public class OpenShiftInternalRuntime extends KubernetesInternalRuntime<OpenShif
     }
 
     return createdRoutes;
-  }
-
-  @Override
-  protected Optional<String> findHostForServicePort(Service service, int port)
-      throws InfrastructureException {
-    Optional<ServicePort> foundPort = Services.findPort(service, port);
-    if (!foundPort.isPresent()) {
-      return Optional.empty();
-    }
-
-    for (Route route : project.routes().get()) {
-      RouteSpec spec = route.getSpec();
-      if (spec.getTo().getName().equals(service.getMetadata().getName())
-          && spec.getPort().getTargetPort().getStrVal().equals(foundPort.get().getName())) {
-        return Optional.of(route.getSpec().getHost());
-      }
-    }
-    return Optional.empty();
   }
 }
